Move ResponseGenerator from generate.py to server.py
ResponseGenerator is server-specific infrastructure (threaded queue,
request dispatching) — it belongs with the server, not the generation
library. BatchGenerator stays in generate.py for offline batch use.
