+++ mlx_vlm/generate.py
@@ -5,15 +5,11 @@ import functools
 import json
 import time
 from dataclasses import dataclass
-from queue import Empty as QueueEmpty
-from queue import Queue
-from threading import Thread
 from typing import (
     Any,
     Callable,
     Dict,
     Generator,
-    Iterator,
     List,
     Optional,
     Tuple,
@@ -1315,542 +1311,6 @@ class BatchGenerator:
     def next(self, **kwargs):
         with mx.stream(generation_stream):
             return self._next(**kwargs)
-
-
-# =============================================================================
-# ResponseGenerator - Concurrent Request Handling with Threaded Batching
-# =============================================================================
-
-
-@dataclass
-class GenerationArguments:
-    """Arguments for a generation request."""
-
-    max_tokens: int = 256
-    temperature: float = 0.0
-    top_p: float = 1.0
-    seed: Optional[int] = None
-    repetition_penalty: Optional[float] = None
-    repetition_context_size: Optional[int] = 20
-    logit_bias: Optional[Dict[int, float]] = None
-
-
-@dataclass
-class GenerationContext:
-    """Context returned when a request is queued."""
-
-    uid: int
-    prompt_tokens: int
-
-
-@dataclass
-class StreamingToken:
-    """A single token response during streaming generation."""
-
-    text: str
-    token: int
-    logprobs: float
-    finish_reason: Optional[str]
-    peak_memory: float = 0.0
-
-
-class ResponseGenerator:
-    """
-    Manages concurrent request handling with threaded batching.
-
-    This class runs a dedicated generation thread that processes requests
-    from a queue using BatchGenerator. Requests are batched when possible
-    (same model, no seed) for better throughput.
-
-    Usage:
-        generator = ResponseGenerator(model, processor, stop_tokens)
-
-        # Submit a request
-        ctx, token_iterator = generator.generate(request, args)
-
-        # Consume tokens
-        for token in token_iterator:
-            print(token.text, end="")
-
-        # Cleanup when done
-        generator.stop_and_join()
-    """
-
-    def __init__(
-        self,
-        model,
-        processor,
-        stop_tokens: Optional[set] = None,
-    ):
-        """
-        Initialize the ResponseGenerator.
-
-        Args:
-            model: The VLM model.
-            processor: The tokenizer/processor.
-            stop_tokens: Set of token IDs that signal end of generation.
-        """
-        self.model = model
-        self.processor = processor
-        self.stop_tokens = stop_tokens or set()
-        self.tokenizer = (
-            processor.tokenizer if hasattr(processor, "tokenizer") else processor
-        )
-        self.requests: Queue = Queue()
-        self._stop = False
-        self._supports_batching = True
-
-        self._generation_thread = Thread(target=self._generate, daemon=True)
-        self._generation_thread.start()
-
-    def stop_and_join(self):
-        """Stop the generation thread gracefully."""
-        self._stop = True
-        # Put a sentinel to wake up the thread if it's waiting
-        self.requests.put(None)
-        self._generation_thread.join(timeout=5.0)
-
-    def generate(
-        self,
-        prompt: str,
-        images: Optional[List] = None,
-        audio: Optional[List] = None,
-        args: Optional[GenerationArguments] = None,
-        progress_callback: Optional[Callable[[int, int], None]] = None,
-    ) -> Tuple[GenerationContext, Iterator[StreamingToken]]:
-        """
-        Submit a generation request and return a context and token iterator.
-
-        Args:
-            prompt: The formatted prompt text.
-            images: Optional list of images (paths, URLs, or PIL images).
-            audio: Optional list of audio files.
-            args: Generation arguments (max_tokens, temperature, etc.).
-            progress_callback: Optional callback for prefill progress.
-
-        Returns:
-            Tuple of (GenerationContext, token iterator).
-        """
-        args = args or GenerationArguments()
-        response_queue: Queue = Queue()
-
-        # Preprocess images/audio BEFORE queueing to avoid blocking generation thread
-        preprocessed_inputs = None
-        if images or audio:
-            add_special_tokens = (
-                not hasattr(self.processor, "chat_template")
-                if self.model.config.model_type in ["gemma3", "gemma3n"]
-                else True
-            )
-            image_token_index = getattr(self.model.config, "image_token_index", None)
-
-            preprocessed_inputs = prepare_inputs(
-                self.processor,
-                images=images,
-                audio=audio,
-                prompts=prompt,
-                image_token_index=image_token_index,
-                add_special_tokens=add_special_tokens,
-            )
-
-        # Create request tuple
-        request = {
-            "prompt": prompt,
-            "images": images,
-            "audio": audio,
-            "args": args,
-            "progress_callback": progress_callback,
-            "preprocessed_inputs": preprocessed_inputs,
-        }
-
-        self.requests.put((response_queue, request))
-
-        # Get the initial context
-        ctx = response_queue.get()
-        if isinstance(ctx, Exception):
-            raise ctx
-
-        def token_iterator():
-            """Iterate over response tokens from the queue."""
-            try:
-                while True:
-                    response = response_queue.get(
-                        timeout=30.0
-                    )  # Timeout to prevent infinite wait
-                    if response is None:
-                        break
-                    if isinstance(response, Exception):
-                        raise response
-                    if isinstance(response, tuple):
-                        # Progress update (processed_tokens, total_tokens)
-                        if progress_callback is not None:
-                            progress_callback(*response)
-                        continue
-                    yield response
-            finally:
-                # Always request cleanup when iterator is done
-                self._request_cleanup(ctx.uid)
-
-        return ctx, token_iterator()
-
-    def _request_cleanup(self, uid: int):
-        """Request cleanup of a UID from the batch (called from client side)."""
-        # Put a cleanup request in the queue
-        self.requests.put(("_cleanup", uid))
-
-    def _is_batchable(self, args: GenerationArguments) -> bool:
-        """Check if request can be batched (no seed = batchable)."""
-        return args.seed is None
-
-    def _make_sampler(self, args: GenerationArguments) -> Optional[Callable]:
-        """Create a sampler function for the given arguments."""
-        if args.temperature == 0:
-            return None  # Use default argmax
-
-        def sampler(logprobs: mx.array) -> mx.array:
-            if args.top_p > 0 and args.top_p < 1.0:
-                return top_p_sampling(logprobs, args.top_p, args.temperature)
-            else:
-                return mx.random.categorical(logprobs * (1 / args.temperature))
-
-        return sampler
-
-    def _preprocess_request(self, request: dict) -> Tuple[mx.array, dict]:
-        """
-        Preprocess a request to get input_ids and embeddings.
-
-        Returns:
-            Tuple of (input_ids, gen_kwargs with embeddings if multimodal).
-        """
-        prompt = request["prompt"]
-        images = request["images"]
-        audio = request["audio"]
-
-        # Use preprocessed inputs if available (preprocessing done before queueing)
-        if request.get("preprocessed_inputs") is not None:
-            inputs = request["preprocessed_inputs"]
-        else:
-            add_special_tokens = (
-                not hasattr(self.processor, "chat_template")
-                if self.model.config.model_type in ["gemma3", "gemma3n"]
-                else True
-            )
-
-            image_token_index = getattr(self.model.config, "image_token_index", None)
-
-            inputs = prepare_inputs(
-                self.processor,
-                images=images,
-                audio=audio,
-                prompts=prompt,
-                image_token_index=image_token_index,
-                add_special_tokens=add_special_tokens,
-            )
-
-        input_ids = inputs.get("input_ids", None)
-        pixel_values = inputs.get("pixel_values", None)
-        mask = inputs.get("attention_mask", None)
-
-        data_kwargs = {
-            k: v
-            for k, v in inputs.items()
-            if k not in ["input_ids", "pixel_values", "attention_mask"]
-        }
-
-        gen_kwargs = {}
-        if pixel_values is not None:
-            embedding_output = self.model.get_input_embeddings(
-                input_ids, pixel_values, mask=mask, **data_kwargs
-            )
-
-            gen_kwargs = {
-                **data_kwargs,
-                **embedding_output.to_dict(),
-            }
-
-        return input_ids, gen_kwargs
-
-    def _serve_single(self, response_queue: Queue, request: dict):
-        """
-        Serve a single non-batchable request (e.g., with seed).
-
-        Uses stream_generate for direct generation.
-        """
-        prompt = request["prompt"]
-        images = request["images"]
-        audio = request["audio"]
-        args = request["args"]
-        preprocessed_inputs = request.get("preprocessed_inputs")
-
-        try:
-            # Set seed if specified
-            if args.seed is not None:
-                mx.random.seed(args.seed)
-
-            # Build kwargs for stream_generate
-            gen_kwargs = {
-                "max_tokens": args.max_tokens,
-                "temperature": args.temperature,
-                "top_p": args.top_p,
-                "repetition_penalty": args.repetition_penalty,
-                "repetition_context_size": args.repetition_context_size,
-                "logit_bias": args.logit_bias,
-            }
-
-            # Use preprocessed inputs if available to skip re-preprocessing
-            if preprocessed_inputs is not None:
-                gen_kwargs.update(preprocessed_inputs)
-                prompt_tokens = preprocessed_inputs.get("input_ids").size
-            else:
-                # Fall back to getting input_ids for context
-                input_ids, _ = self._preprocess_request(request)
-                prompt_tokens = (
-                    input_ids.size if hasattr(input_ids, "size") else len(input_ids)
-                )
-
-            # Send context
-            ctx = GenerationContext(uid=-1, prompt_tokens=prompt_tokens)
-            response_queue.put(ctx)
-
-            # Generate using stream_generate
-            for result in stream_generate(
-                self.model,
-                self.processor,
-                prompt,
-                image=images if preprocessed_inputs is None else None,
-                audio=audio if preprocessed_inputs is None else None,
-                **gen_kwargs,
-            ):
-                # Handle token as scalar or array
-                tok = result.token
-                if hasattr(tok, "item"):
-                    tok = tok.item()
-
-                # Safely extract logprobs
-                lp = 0.0
-                if result.logprobs is not None:
-                    try:
-                        lp_val = result.logprobs[tok]
-                        if hasattr(lp_val, "item"):
-                            lp = lp_val.item()
-                        else:
-                            lp = float(lp_val)
-                    except (IndexError, TypeError, ValueError):
-                        lp = 0.0
-
-                token = StreamingToken(
-                    text=result.text,
-                    token=tok,
-                    logprobs=lp,
-                    finish_reason=None,
-                    peak_memory=result.peak_memory,
-                )
-                response_queue.put(token)
-
-            # Send final token with finish reason
-            response_queue.put(None)
-
-        except Exception as e:
-            response_queue.put(e)
-        finally:
-            mx.clear_cache()
-
-    def _generate(self):
-        """
-        Main generation loop running in a dedicated thread.
-
-        Processes requests with continuous batching - multiple requests
-        can be processed together for better throughput.
-        """
-        batch_generator = None
-        # Maps UID -> {rqueue, tokens, gen_kwargs}
-        batch_results = {}
-
-        while not self._stop:
-            try:
-                # Determine timeout based on whether we have active work
-                timeout = 0.01 if batch_results else 0.1
-
-                # Try to get a new request
-                try:
-                    item = self.requests.get(timeout=timeout)
-                except QueueEmpty:
-                    item = None
-
-                if item is None:
-                    if self._stop:
-                        break
-                    # No new request - process existing batch if any
-                    if batch_generator is not None and batch_results:
-                        self._process_batch_step(batch_generator, batch_results)
-                    continue
-
-                # Sentinel for shutdown
-                if item is None:
-                    break
-
-                # Handle cleanup request
-                if isinstance(item, tuple) and len(item) == 2 and item[0] == "_cleanup":
-                    cleanup_uid = item[1]
-                    if cleanup_uid in batch_results:
-                        del batch_results[cleanup_uid]
-                    if batch_generator is not None:
-                        batch_generator.remove([cleanup_uid])
-                    continue
-
-                response_queue, request = item
-                args = request["args"]
-
-                # Non-batchable: model doesn't support batching or has seed
-                # Note: image/audio requests now go through batch generator for interleaved processing
-                if not self._supports_batching or not self._is_batchable(args):
-                    # Drain current batch first
-                    if batch_generator is not None and batch_results:
-                        while batch_results:
-                            self._process_batch_step(batch_generator, batch_results)
-                        batch_generator.close()
-                        batch_generator = None
-
-                    self._serve_single(response_queue, request)
-                    continue
-
-                # Initialize batch generator if needed
-                if batch_generator is None:
-                    batch_generator = BatchGenerator(
-                        self.model.language_model,
-                        self.processor,
-                        stop_tokens=self.stop_tokens,
-                    )
-
-                # Preprocess and add to batch
-                try:
-                    input_ids, gen_kwargs = self._preprocess_request(request)
-                    prompt_tokens = (
-                        input_ids.size if hasattr(input_ids, "size") else len(input_ids)
-                    )
-
-                    sampler = self._make_sampler(args)
-                    has_embeddings = bool(gen_kwargs.get("inputs_embeds") is not None)
-
-                    # Drain pending text-only prompts BEFORE inserting an
-                    # image request, so they don't get prefilled without
-                    # the correct embeddings.
-                    if has_embeddings and batch_generator.unprocessed_prompts:
-                        while batch_generator.unprocessed_prompts:
-                            drain_responses = batch_generator.next()
-                            if drain_responses:
-                                self._handle_responses(
-                                    drain_responses, batch_results
-                                )
-                            else:
-                                break
-
-                    # Insert and get UID
-                    (uid,) = batch_generator.insert(
-                        [input_ids.squeeze(0).tolist()],
-                        max_tokens=args.max_tokens,
-                        samplers=[sampler],
-                    )
-
-                    # Send context to client
-                    ctx = GenerationContext(uid=uid, prompt_tokens=prompt_tokens)
-                    response_queue.put(ctx)
-
-                    # Track this request by UID
-                    batch_results[uid] = {
-                        "rqueue": response_queue,
-                        "tokens": [],
-                        "gen_kwargs": gen_kwargs,
-                        "needs_prefill": has_embeddings,
-                    }
-
-                    # Prefill this image request immediately with its embeddings
-                    if has_embeddings:
-                        responses = batch_generator.next(**gen_kwargs)
-                        if responses:
-                            self._handle_responses(responses, batch_results)
-
-                except Exception as e:
-                    response_queue.put(e)
-
-            except Exception as e:
-                print(f"Error in generation thread: {e}")
-                import traceback
-
-                traceback.print_exc()
-
-        # Cleanup
-        if batch_generator is not None:
-            batch_generator.close()
-
-    def _handle_responses(self, responses, batch_results: dict):
-        """Handle responses from batch generator and dispatch to clients."""
-        for r in responses:
-            # r.uid identifies which request this token belongs to
-            if r.uid not in batch_results:
-                continue
-
-            result = batch_results[r.uid]
-            rqueue = result["rqueue"]
-            tokens = result["tokens"]
-
-            # Don't include stop tokens in output
-            if r.finish_reason == "stop":
-                text = ""
-            else:
-                # Collect token for this specific request
-                tokens.append(r.token)
-
-                # Decode incrementally for this request's tokens only
-                if len(tokens) >= 2:
-                    prev_text = self.tokenizer.decode(tokens[:-1])
-                    curr_text = self.tokenizer.decode(tokens)
-                    text = curr_text[len(prev_text) :]
-                else:
-                    text = self.tokenizer.decode(tokens)
-
-            # Handle token as scalar or array
-            tok = r.token
-            if hasattr(tok, "item"):
-                tok = tok.item()
-
-            # Safely extract logprobs
-            lp = 0.0
-            if r.logprobs is not None:
-                try:
-                    lp_val = r.logprobs[tok]
-                    if hasattr(lp_val, "item"):
-                        lp = lp_val.item()
-                    else:
-                        lp = float(lp_val)
-                except (IndexError, TypeError, ValueError):
-                    lp = 0.0
-
-            # Send token to this request's queue
-            token_response = StreamingToken(
-                text=text,
-                token=tok,
-                logprobs=lp,
-                finish_reason=r.finish_reason,
-                peak_memory=r.peak_memory,
-            )
-            rqueue.put(token_response)
-
-            # Clean up finished request
-            if r.finish_reason is not None:
-                rqueue.put(None)  # Signal end
-                del batch_results[r.uid]
-
-    def _process_batch_step(self, batch_generator: BatchGenerator, batch_results: dict):
-        """Process one generation step for all active requests."""
-        # Don't pass any gen_kwargs during batch step - embeddings were already
-        # processed during individual prefills. Passing image kwargs here would
-        # cause shape mismatches with text-only requests.
-        responses = batch_generator.next()
-        if not responses:
-            return
-
-        self._handle_responses(responses, batch_results)
 
 
 def batch_generate(
